Move special operators module to utils (#231)
* move _special_operators.py from hamiltonians to utils

* move test file

* fix import statements

* lint (trying to trigger travis)
