+++ src/openfermion/hamiltonians/__init__.py
@@ -30,11 +30,6 @@ from ._plane_wave_hamiltonian import (dual_basis_external_potential,
                                       jordan_wigner_dual_basis_hamiltonian,
                                       wigner_seitz_length_scale)
 
-from ._special_operators import (up_index, down_index,
-                                 majorana_operator, number_operator,
-                                 s_squared_operator, s_plus_operator,
-                                 s_minus_operator, sz_operator)
-
 # Imports out of alphabetical order to avoid circular dependancy.
 
 from ._hubbard import fermi_hubbard
