+++ src/openfermion/ops/_quadratic_hamiltonian_test.py
@@ -17,12 +17,11 @@ import numpy
 import unittest
 
 from openfermion.config import EQ_TOLERANCE
-from openfermion.hamiltonians import majorana_operator
 from openfermion.ops import (normal_ordered,
                              FermionOperator, QuadraticHamiltonian)
 from openfermion.ops._quadratic_hamiltonian import antisymmetric_canonical_form
 from openfermion.transforms import get_fermion_operator, get_sparse_operator
-from openfermion.utils import get_ground_state
+from openfermion.utils import get_ground_state, majorana_operator
 from openfermion.utils._testing_utils import (random_antisymmetric_matrix,
                                               random_hermitian_matrix)
