+++ src/openfermion/transforms/_bravyi_kitaev_test.py
@@ -16,13 +16,12 @@ from __future__ import absolute_import
 import numpy
 import unittest
 
-from openfermion.hamiltonians import number_operator
 from openfermion.ops import (FermionOperator,
                              QubitOperator)
 from openfermion.transforms import (bravyi_kitaev,
                                     get_sparse_operator,
                                     jordan_wigner)
-from openfermion.utils import eigenspectrum
+from openfermion.utils import eigenspectrum, number_operator
 
 
 class BravyiKitaevTransformTest(unittest.TestCase):
