+++ src/openfermion/transforms/_conversion_test.py
@@ -17,7 +17,6 @@ import copy
 import numpy
 import unittest
 
-from openfermion.hamiltonians import number_operator
 from openfermion.ops import (FermionOperator, InteractionOperator,
                              normal_ordered, QubitOperator)
 from openfermion.ops._interaction_operator import InteractionOperatorError
