+++ src/openfermion/utils/_special_operators.py
